% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aids.id.R
\docType{data}
\name{aids.id}
\alias{aids.id}
\title{AIDS data-set - baseline.}
\format{A data frame with 467 rows and 9 variables:
\describe{
  \item{patient}{id}
  \item{Time}{survival time}
  \item{death}{died? 1 = yes}
  \item{CD4}{biomarker}
  \item{obstime}{follow-up time}
  \item{drug}{drug group}
  \item{gender}{sex}
  \item{prevOI}{previously had AIDS?}
  \item{AZT}{intolerant to AZT?}
}}
\usage{
aids.id
}
\description{
A dataset containing data for 467 patients
}
\keyword{datasets}
