% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_jm.R
\name{fit_jm}
\alias{fit_jm}
\title{Bayesian inference for joint models}
\usage{
fit_jm(fixed_long, random_long, fixed_surv, data_long, data_surv, id_long,
  id_surv, model, timeVar, bh_nknots = 2, spline_tv, Q = 15,
  priors = list(), ...)
}
\arguments{
\item{fixed_long}{A two-sided formula for fixed effects of longitudinal sub-model}

\item{random_long}{A one-sided formula for random effects of longitudinal sub-model}

\item{fixed_surv}{A two-sided formula for survival time, event indicator and 
baseline covariates to be included in survival sub-model, 
e.g. \code{cbind(Time, death) ~ x}}

\item{data_long}{A data frame that includes the longitudinal data}

\item{data_surv}{A data frame that includes the survival data}

\item{id_long}{A character string that indicates the column name for the id column in \code{data_long}}

\item{id_surv}{A character string that indicates the column name for the id column in \code{data_surv}}

\item{model}{A character string for model identification; options are: 
"nor_nor", "t_t_mod1", "t_t_mod2", "t_t_mod3", "nor_t_mod3" (not yet available), 
"t_t_tv", "nor_t_tv" (not yet available)}

\item{timeVar}{A character string for the column name of the time variable in \code{data_long}}

\item{bh_nknots}{A numeric value for the number of knots for baseline hazard b-spline specification; default is 2}

\item{spline_tv}{A list with two elements; first element is the name of the time variable, 
and number of knots}

\item{Q}{A numeric value for the number of quadratures; default is 15}

\item{priors}{A list of hyperparameters; theta, Omega, sigma_B, sigma_Z, beta (for tv), zeta, omega, eta. 
See details below.}

\item{...}{to be passed to the \code{stan} function}
}
\value{
Returns the output of the \code{stan} function
}
\description{
Fits jont models with Normal and non-Normal distributions
}
\details{
This is a wrapper function for fitting mixed effects models. 
         Cauchy distribution is assumed as the prior for theta (QR decomposed alpha), 
         half-Cauchy for sigma_B (standard deviations of var-cov of B), 
         half-Cauchy for sigma_Z (standard deviation of error),
         Cauchy for beta (time-varying degree of freedom parameters),
         Cauchy for zeta (baseline hazard spline coefficients),
         Cauchy for omega (survival sub-model regression coefficients),
         Cauchy for eta (association parameter)
}
\examples{
For examples, see the repository at https://github.com/ozgurasarstat/robjm-run                                              
}
