% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{in_interval}
\alias{in_interval}
\title{CreateS a vector indicator values}
\usage{
in_interval(x, knots)
}
\arguments{
\item{x}{A numeric value for the time variable}

\item{knots}{A numeric vector that includes the knot locations}
}
\description{
Creates a vector of zeros and one, one meaning the value being included in the interval
}
