% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orthodont.R
\docType{data}
\name{Orthodont}
\alias{Orthodont}
\title{Orthodont data-set.}
\format{A data frame with 108 rows and 4 variables:
\describe{
  \item{distance}{outcome}
  \item{age}{age in years}
  \item{Subject}{id}
  \item{Sex}{sex of the child}
}}
\usage{
Orthodont
}
\description{
A dataset containing dental data for 27 children at ages 8, 10, 12, 14
}
\keyword{datasets}
