% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSurv_jm.R
\name{predSurv_jm}
\alias{predSurv_jm}
\title{Predict survival probabilities}
\usage{
predSurv_jm(object, newdata, l, h, inc, model, ...)
}
\arguments{
\item{object}{A fitted object from \code{fit_jm}}

\item{newdata}{}
}
\description{
A function to predict survival probabilities
}
