% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{prob_summary}
\alias{prob_summary}
\title{Function to summarise the probability estimates}
\usage{
prob_summary(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Function to summarise the probability estimates
}
