% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pw_mat}
\alias{pw_mat}
\title{Creates a matrix of indicator values}
\usage{
pw_mat(x, knots)
}
\arguments{
\item{x}{A numeric vector for the time variable}

\item{knots}{A numeric vector that includes the knot locations}
}
\description{
Creates a matrix of zeros and ones, ones meaning being included in the interval
}
