% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_prob_calc.R
\name{surv_prob_calc}
\alias{surv_prob_calc}
\title{Calculate survival probability}
\usage{
surv_prob_calc(t, x, d, c, timeVar, log_lambda, log_nu, omega, eta, alpha,
  B, wt, pt, Q, bh)
}
\arguments{
\item{t}{scalar beyond which survival probabilities to be calculated}

\item{x}{design matrix of the fixed effects}

\item{d}{design matrix of the random effects}

\item{c}{design matrix of the survival sub-model}

\item{log_lambda}{matrix of posterior draws of log_lambda}

\item{log_nu}{matrix of posterior draws of log_nu}

\item{omega}{matrix of posterior draws of omega}

\item{eta}{matrix of posterior draws of eta}

\item{alpha}{matrix of posterior draws of alpha}

\item{B}{list of matrices for posterior draws of B}

\item{wt}{vector of weights}

\item{pt}{vector of abscissa}

\item{Q}{scalar for number of quadratures}
}
\description{
A unction to calculate survival probability for a single subject beyond a 
             single time point for a single posterior draw of the parameters
}
